requirements.txt
Flask==2.1.2
PyJWT==2.4.0
PyYAML==6.0
requests==2.27.1
